Provide full hid, usbhid and uhid modules instead of a glue to the current hid
Kernels <2.6.38 are incompatible with some of the multitouch devices
which present a lot of fields in their report desriptor.
Instead of maintaining a glue to the current hid subsystem, just ship a
new and up to date one.

The code of hid-core.c hid-debug.c hid-input.c hid-multitouch.c hidraw.c
uhid.c and usbhid/* is synchronised with upstream.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>