Add usbhid from 2.6.34 for earlier kernels
In 2.6.38 (at least) there has been changes regarding the usb
implementation. This is not the point of this backport, so found
a way for earlier kernels to compile.

Taking the 2.6.34 version as this is the last stable release before
the change.

Note that hiddev is removed as it is of no use here.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>